refactor: merge legacy files into legacy_xlayer.rs
