+++ crates/rpc/rpc-builder/src/lib.rs
@@ -98,7 +98,7 @@ pub use eth::EthHandlers;
 
 // Rpc server metrics
 mod metrics;
-mod xlayer_legacy;
+mod legacy_xlayer;
 use crate::middleware::RethRpcMiddleware;
 pub use metrics::{MeteredRequestFuture, RpcRequestMetricsService};
 use reth_chain_state::CanonStateSubscriptions;
