+++ crates/rpc/rpc-eth-types/src/lib.rs
@@ -15,7 +15,7 @@ pub mod error;
 pub mod fee_history;
 pub mod gas_oracle;
 pub mod id_provider;
-pub mod legacy;
+pub mod legacy_xlayer;
 pub mod logs_utils;
 pub mod pending_block;
 pub mod receipt;
@@ -35,7 +35,7 @@ pub use gas_oracle::{
     GasCap, GasPriceOracle, GasPriceOracleConfig, GasPriceOracleResult, RPC_DEFAULT_GAS_CAP,
 };
 pub use id_provider::EthSubscriptionIdProvider;
-pub use legacy::{LegacyRpcClient, LegacyRpcConfig};
+pub use legacy_xlayer::{LegacyRpcClient, LegacyRpcConfig};
 pub use pending_block::{PendingBlock, PendingBlockEnv, PendingBlockEnvOrigin};
 pub use transaction::TransactionSource;
 pub use tx_forward::ForwardConfig;
